Cleanup
* Nsd with internal functionallity
* NsdServiceDiscovery as interface
* extension on Context.getNetworkServiceDiscovery, that uses the internal NetworkServiceDiscoveryImpl
* more separated
