+++ app/src/main/java/de/abacuselectronics/aroiorc/extensions/Float.kt
@@ -1,4 +1,4 @@
-package de.abacuselectronics.aroiorc.aroioktx
+package de.abacuselectronics.aroiorc.extensions
 
 import android.content.Context
 import android.util.DisplayMetrics
