+++ app/src/main/java/de/abacuselectronics/aroiorc/extensions/FragmentTransaction.kt
@@ -1,4 +1,4 @@
-package de.abacuselectronics.aroiorc.aroioktx
+package de.abacuselectronics.aroiorc.extensions
 
 import androidx.fragment.app.FragmentTransaction
 import de.abacuselectronics.aroiorc.R
