+++ app/src/main/java/de/abacuselectronics/aroiorc/extensions/TextView.kt
@@ -1,4 +1,4 @@
-package de.abacuselectronics.aroiorc.aroioktx
+package de.abacuselectronics.aroiorc.extensions
 
 import android.os.Build.VERSION.SDK_INT
 import android.os.Build.VERSION_CODES.M
