+++ app/src/main/java/de/abacuselectronics/aroiorc/ui/custom/OneLineSwitchView.kt
@@ -11,7 +11,7 @@ import androidx.core.content.ContextCompat
 import androidx.core.content.withStyledAttributes
 import com.google.android.material.switchmaterial.SwitchMaterial
 import de.abacuselectronics.aroiorc.R
-import de.abacuselectronics.aroiorc.aroioktx.textAppearance
+import de.abacuselectronics.aroiorc.extensions.textAppearance
 
 class OneLineSwitchView @JvmOverloads constructor(
   context: Context,
