+++ app/src/main/java/de/abacuselectronics/aroiorc/ui/recycler/items/OverlineItem.kt
@@ -6,7 +6,7 @@ import android.widget.LinearLayout
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.recyclerview.widget.RecyclerView
 import de.abacuselectronics.aroiorc.R
-import de.abacuselectronics.aroiorc.aroioktx.updateMarginsRelative
+import de.abacuselectronics.aroiorc.extensions.updateMarginsRelative
 import de.abacuselectronics.aroiorc.ui.recycler.RecyclerItem
 
 class OverlineItem(
