+++ app/src/main/java/de/abacuselectronics/aroiorc/viewmodel/ViewModelFactory.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import de.abacuselectronics.aroiorc.ui.start.AroioListViewModel
-import de.lennartegb.nsd.NetworkServiceDiscovery
+import de.lennartegb.nsd.extensions.getNetworkServiceDiscovery
 
 class ViewModelFactory(private val context: Context) :
 	ViewModelProvider.Factory {
@@ -13,7 +13,7 @@ class ViewModelFactory(private val context: Context) :
 	override fun <T : ViewModel?> create(modelClass: Class<T>): T {
 		when {
 			modelClass.isAssignableFrom(AroioListViewModel::class.java) -> {
-				val nsd = NetworkServiceDiscovery(context)
+				val nsd = context.getNetworkServiceDiscovery()
 				return AroioListViewModel(nsd) as T
 			}
 			else                                                        -> {
