+++ nsd/src/main/java/de/lennartegb/nsd/NsdResult.kt
@@ -1,0 +1,8 @@
+package de.lennartegb.nsd
+
+import android.net.nsd.NsdServiceInfo
+
+sealed class NsdResult {
+	class ServiceFound(val service: NsdServiceInfo) : NsdResult()
+	class ServiceLost(val service: NsdServiceInfo) : NsdResult()
+}
