+++ nsd/src/main/java/de/lennartegb/nsd/extensions/Context.kt
@@ -1,0 +1,8 @@
+package de.lennartegb.nsd.extensions
+
+import android.content.Context
+import de.lennartegb.nsd.NetworkServiceDiscovery
+import de.lennartegb.nsd.NetworkServiceDiscoveryImpl
+
+fun Context.getNetworkServiceDiscovery(): NetworkServiceDiscovery =
+	NetworkServiceDiscoveryImpl(this)
